Replaced the monolithic script by atomic prefabs w script.
