+++ GameJam2022/Assets/Prefabs/FakeLightningTarget.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: dab94dd7d40caa342a05c9694e1198ce
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
