+++ GameJam2022/Assets/Prefabs/weapons.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d0b8dd3a414dee94cadc7e2fb3e85e04
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
