+++ GameJam2022/Assets/Prefabs/weapons/FlamethrowerType.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b7f2d83bc6ee2f24b9a54b22d287364f
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
