+++ GameJam2022/Assets/Prefabs/weapons/GrenadeLauncherType.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: dbb6cf50c9771c44d9a455f398d214e7
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
