+++ GameJam2022/Assets/Prefabs/weapons/LightningType.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3303a5cb19661ea4584d83a340c7fb8a
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
