+++ GameJam2022/Assets/Prefabs/weapons/PistolType.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d223ace760dbd884aae633e9d3ba9cd2
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
