+++ GameJam2022/Assets/Scripts/FakeTargetBehavior.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b060fb1b243a0074283df480cdb99d54
+guid: 3acfbf0ba3da0934e949269dc3427632
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
