+++ GameJam2022/Assets/Scripts/FlamethrowerBehavior.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: daed49fce8720d542a33e89bced556db
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
