+++ GameJam2022/Assets/Scripts/GrenadeLauncherBehavior.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8dec720a0c3be3843bb6b4ccdd2537b2
+guid: 1ca186eb1e638884aa2594f9bc5c664d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
