+++ GameJam2022/Assets/Scripts/IShooting.cs
@@ -1,0 +1,4 @@
+public interface IShooting
+{
+    public void Shoot();
+}
