+++ GameJam2022/Assets/Scripts/IShooting.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f1c18dc3fab70cd4f9f4ba2a2fe8982a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
