+++ GameJam2022/Assets/Scripts/PistolBehavior.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 09f3cf2fa3c9efd4292e5edf1931018d
+guid: ddff848e18da27a49969173fd51b17a7
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
